+++ client/scripts/ui/emu-debug-ui.js
@@ -47,7 +47,7 @@ colorLut.set('WHITE', 'rgba(255,255,255,');
 colorLut.set('GREEN', 'rgba(0,255,0,');
 colorLut.set('BLACK', 'rgba(0,0,0,0)');
 
-function updateCanvas(emuState, cpuRunningState, bleState) {
+function updateCanvas(emuState, cpuRunningState, audioState, bleState) {
   if (!emuState) {
     return;
   }
@@ -75,11 +75,13 @@ function updateCanvas(emuState, cpuRunningState, bleState) {
     LEFT_X_OFFSET, YPOS_GENERIC_DATA + 120);
   canvas.fillText('SND WRITE (DATA/CTL): ' + emuState.asic.sound.writeDataBytes + '/' + emuState.asic.sound.writeControlBytes,
     LEFT_X_OFFSET, YPOS_GENERIC_DATA + 130);
-  canvas.fillText('EMU TIME: ' + emuState.asic.wpc.time, LEFT_X_OFFSET, YPOS_GENERIC_DATA + 140);
+  canvas.fillText('SND STATE: ' + audioState, LEFT_X_OFFSET, YPOS_GENERIC_DATA + 140);
+  canvas.fillText('EMU TIME: ' + emuState.asic.wpc.time, LEFT_X_OFFSET, YPOS_GENERIC_DATA + 150);
   if (bleState) {
-    canvas.fillText('BLE MESSAGES: ' + bleState.bleMessageCount, LEFT_X_OFFSET, YPOS_GENERIC_DATA + 150);
-    canvas.fillText('PIN UPTIME S: ' + bleState.pinballUptimeS, LEFT_X_OFFSET, YPOS_GENERIC_DATA + 160);
+    canvas.fillText('BLE MESSAGES: ' + bleState.bleMessageCount, LEFT_X_OFFSET, YPOS_GENERIC_DATA + 160);
+    canvas.fillText('PIN UPTIME S: ' + bleState.pinballUptimeS, LEFT_X_OFFSET, YPOS_GENERIC_DATA + 170);
   }
+
   canvas.fillText('DMD PAGE MAP: ' + emuState.asic.dmd.dmdPageMapping, LEFT_X_OFFSET, YPOS_DMD_DATA + 10);
   canvas.fillText('DMD ACTIVE PAGE: ' + emuState.asic.dmd.activepage, LEFT_X_OFFSET, YPOS_DMD_DATA + 20);